Optimize cache for reentrant tools
The cache digest can be supplied by the caller as a dispatch option.
The cache digest is stored in the tool options.
A new :pre-dispatch callback is passed the tool options, include cache digest.

This is somewhat specific to nucli, which has some commands that are 'reentrant'; a Bash script calls back into the Babashka wrapper, and recomputing the cache digest is unnecessary.